+++ src/auth/auth.guard.ts
@@ -1,26 +1,20 @@
-import { Injectable, CanActivate, ExecutionContext, Inject } from '@nestjs/common';
+import { CanActivate, ExecutionContext, Inject, Injectable, SetMetadata } from '@nestjs/common';
 import { Reflector } from '@nestjs/core';
 import * as Logger from 'bunyan';
 import { Observable } from 'rxjs';
-import { Request } from 'express';
-import { ReflectMetadata } from '@nestjs/common';
 import { createLogger } from '../gcloud/logging';
-import { Context, isContext } from '../datastore/context';
+import { isContext } from '../datastore/context';
 import { Configuration, IUser } from '../index';
 import { CONFIGURATION } from '../configuration';
+import { getRequestFromExecutionContext } from '../util';
 
 const logger = createLogger('auth-guard');
 
-interface RequestWithSessionAndDatastore extends Request {
-  session: any;
-  context: Context;
-}
-
-export const Roles = (...roles: string[]) => ReflectMetadata('roles', roles);
-export const AllowAnonymous = () => ReflectMetadata('allowAnonymous', true);
-export const Task = () => ReflectMetadata('secure-header', 'x-appengine-taskname');
-export const Cron = () => ReflectMetadata('secure-header', 'x-appengine-cron');
-export const System = () => ReflectMetadata('system', true);
+export const Roles = (...roles: string[]) => SetMetadata('roles', roles);
+export const AllowAnonymous = () => SetMetadata('allowAnonymous', true);
+export const Task = () => SetMetadata('secure-header', 'x-appengine-taskname');
+export const Cron = () => SetMetadata('secure-header', 'x-appengine-cron');
+export const System = () => SetMetadata('system', true);
 
 const reflectValue = <T>(reflector: Reflector, key: string, context: ExecutionContext, defaultValue: T) => {
   const methodValue = reflector.get<T>(key, context.getHandler());
@@ -68,7 +62,7 @@ function isSystemCall(reflector: Reflector, context: ExecutionContext): boolean 
 async function isAuthorizedSystemCall(reflector: Reflector, context: ExecutionContext, secret: Buffer) {
   const { verify } = await import('jsonwebtoken');
 
-  const { headers } = context.switchToHttp().getRequest<Request>();
+  const { headers } = getRequestFromExecutionContext(context);
 
   if (!headers.authorization) {
     return false;
@@ -97,7 +91,7 @@ async function isAuthorizedSystemCall(reflector: Reflector, context: ExecutionCo
 }
 
 function hasSecureHeader(reflector: Reflector, context: ExecutionContext): boolean {
-  const { headers } = context.switchToHttp().getRequest<RequestWithSessionAndDatastore>();
+  const { headers } = getRequestFromExecutionContext(context);
 
   if (!headers) {
     return false;
@@ -113,7 +107,7 @@ function hasSecureHeader(reflector: Reflector, context: ExecutionContext): boole
 }
 
 function getUser(context: ExecutionContext): IUser | undefined {
-  const request = context.switchToHttp().getRequest<RequestWithSessionAndDatastore>();
+  const request = getRequestFromExecutionContext(context);
 
   if (request.context && request.context.user) {
     return request.context.user;