+++ src/module.ts
@@ -14,7 +14,6 @@ import { Configuration } from './configuration';
 import { DatastoreProvider } from './datastore/datastore.provider';
 import { NotFoundFilter } from './filter';
 import { StorageProvider } from './gcloud/storage.provider';
-import { GraphQLMiddleware } from './graphql/GraphQLMiddleware';
 import { ContextMiddleware } from './interceptor';
 import { GmailConfigurer } from './mail/gmail/gmail.configurer';
 import { GmailController } from './mail/gmail/gmail.controller';
@@ -24,6 +23,8 @@ import { MailDiverter } from './mail/mail.diverter';
 import { MailLoggingSender } from './mail/mail-logging.sender';
 import { MAIL_SENDER, MailSender } from './mail/mail.sender';
 import { SearchService } from './search/search.service';
+import { GraphQLDateTime, GraphQLTime } from 'graphql-iso-date';
+import * as _ from 'lodash';
 import { MailWhitelistSender } from './mail/mail-whitelist.sender';
 import { MailSubjectSender } from './mail/mail-subject.sender';
 
@@ -90,7 +91,6 @@ export interface Options {
       provide: APP_GUARD,
       useClass: AuthGuard,
     },
-    GraphQLMiddleware,
   ],
   exports: [
     StorageProvider,
@@ -107,17 +107,26 @@ export interface Options {
   controllers: [AuthController, GmailController],
 })
 export class GCloudModule implements NestModule {
-  constructor(private readonly graphqlConfigurer: GraphQLMiddleware) {}
-
   configure(consumer: MiddlewareConsumer) {
     consumer.apply(ContextMiddleware).forRoutes('*');
-    consumer.apply(GraphQLMiddleware).forRoutes('/api/graphql');
   }
 
   static forConfiguration(options: Options) {
     return {
       module: GCloudModule,
-      imports: [options.configurationModule, options.userModule, GraphQLModule],
+      imports: [
+        options.configurationModule,
+        options.userModule,
+        GraphQLModule.forRoot({
+          path: '/api/graphql',
+          context: (props: any) => _.get(props.req, 'context'),
+          typePaths: ['./src/**/*.graphqls', './node_modules/@3wks/gae-node-nestjs/dist/**/*.graphqls'],
+          resolvers: {
+            Time: GraphQLTime,
+            DateAndTime: GraphQLDateTime,
+          },
+        }),
+      ],
     };
   }
 }