+++ typings/custom-types.d.ts
@@ -1,6 +1,5 @@
 declare module '@google-cloud/connect-datastore';
 declare module '@google-cloud/tasks';
 declare module 'helmet-csp';
-declare module 'merge-graphql-schemas';
 declare module 'passport-google-oauth20';
 declare module 'passport-openidconnect';