+++ CHANGELOG.md
@@ -1,3 +1,43 @@
+## 6.0.0-rc1 (2020-03-00)
+
+- Update to NestJS 6
+
+### Breaking changes:
+
+Lots of changes:
+- Follow official migration guide: [https://docs.nestjs.com/migration-guide]
+- Add @nestjs/platform-express
+```
+npm install @nestjs/platform-express
+```
+- Update `bootstrap.ts` to use express platform
+```
+const app = await NestFactory.create(AppModule, new ExpressAdapter(expressApp), {
+  logger: new BunyanLogger(),
+});
+```
+
+-- In middleware and interceptors use new util method to access Request.
+
+Change:
+```
+const request = context.switchToHttp().getRequest();
+```
+To:
+```
+const request = getRequestFromExecutionContext(context);
+```
+
+-- Add `@ResolveProperty()`, `@Parent` and `@GQLContext` annotations to any graphql property resolver methods
+e.g.
+```
+ @ResolveProperty('recipients')
+  async recipients(
+    @Parent() parent: Message,
+    @GqlContext() context: Context,
+  ): Promise<Recipient[]> {}
+```
+
 ## 5.1.2 (2020-01-31)
 
 - Fix typings to support multiple query predicates for arrays of primitives and union types.